pyproject.toml
[project]
name = "cosmos_env"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "~=3.13.0"
dependencies = [
    "accelerate==0.29.3",
    "bert-score>=0.3.13",
    "config>=0.5.1",
    "datasets>=4.1.1",
    "evaluate>=0.4.6",
    "huggingface-hub>=0.35.1",
    "hydra-core==1.3.2",
    "hydra-submitit-launcher>=1.2.0",
    "mauve-text>=0.4.0",
    "ml-collections>=1.1.0",
    "nltk>=3.9.1",
    "numpy>=2.3.3",
    "omegaconf==2.3.0",
    "pip>=25.2",
    "rouge>=1.0.1",
    "rouge-score>=0.1.2",
    "setuptools>=80.9.0",
    "spacy>=3.8.7",
    "timm==1.0.9",
    "torch>=2.8.0",
    "torch-ema>=0.3",
    "torch-optimi>=0.3.2",
    "tqdm>=4.67.1",
    "transformers==4.48.0",
    "wandb>=0.22.0",
    "zstandard==0.25.0"
]